refactor : Piece 리팩토링
1. 인터페이스, 추상클래스 삭제
2. 각 말들의 중복로직을 Piece라는 부모클래스에서 구현
3. Pawn과 Knight만 move메서드 재정의
