+++ README.md
@@ -9,9 +9,10 @@
  - DataBaseConnector sout 처리
  - BoardCreator 메서드 명 동사 처리
  - ChessGame play 고려
- - Abstract 중복 로직 추상화
  - Direction enum
- - isAlly 캐스팅 미필요 처리
  - 커넥션 실패에 대한 별도 예외 처리
 
 ##Done
+ - isAlly 캐스팅 미필요 처리
+ - Abstract 중복 로직 추상화
+
