+++ src/main/java/chess/domain/pieces/Piece.java
@@ -4,16 +4,66 @@ import chess.domain.ChessTeam;
 import chess.domain.Direction;
 import chess.domain.Point;
 
-public interface Piece {
-    Direction move(Point p1, Point p2);
+import java.util.List;
+import java.util.Objects;
 
-    Direction attack(Point p1, Point p2);
+public class Piece {
+    private final ChessTeam team;
+    private final List<Direction> directions;
+    private PieceInfo info;
 
-    boolean isAlly(Piece piece);
+    Piece(ChessTeam team, PieceInfo info, List<Direction> directions) {
+        this.team = team;
+        this.info = info;
+        this.directions = directions;
+    }
 
-    boolean isTurn(ChessTeam team);
+    public boolean isKing() {
+        return false;
+    }
 
-    boolean isKing();
+    public Direction move(Point p1, Point p2) {
+        Direction vector = p1.direction(p2).vector();
+        if (directions.contains(vector)) {
+            return vector;
+        }
+        throw new IllegalArgumentException();
+    }
 
-    double score();
+    public Direction attack(Point p1, Point p2) {
+        return move(p1, p2);
+    }
+
+    public boolean isAlly(Piece piece) {
+        if (piece == null) throw new IllegalArgumentException();
+        return this.team == piece.team;
+    }
+
+    public boolean isTurn(ChessTeam team) {
+        return this.team == team;
+    }
+
+    public double score() {
+        return info.score();
+    }
+
+    @Override
+    public String toString() {
+        return team.color() + "," + info.toString();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Piece piece = (Piece) o;
+        return team == piece.team &&
+                Objects.equals(directions, piece.directions) &&
+                info == piece.info;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(team, directions, info);
+    }
 }
